requirements.txt
python-telegram-bot==20.4
rapidfuzz
huggingface_hub
